Moved all methods to Helper class
